% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDatasets.R
\docType{data}
\name{dataframeData}
\alias{dataframeData}
\title{Site coordinates (data.frame)}
\format{
An object of class ...
}
\usage{
data(dataframeData)
}
\description{
A data.frame containing coordinates (as longitude and longitude, WGS84) of sites sampled for the marine species Laminaria ochroleuca.
}
