% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnectivityEvents.R
\name{getConnectivityEvents}
\alias{getConnectivityEvents}
\title{Retrieve Oceanographic Connectivity Events by Criteria}
\usage{
getConnectivityEvents(
  hexagonID,
  year = NULL,
  month = NULL,
  day = NULL,
  period = 180
)
}
\arguments{
\item{hexagonID}{A vector of hexagon IDs to filter events.}

\item{year}{(Optional) A numeric vector of years to limit the results.}

\item{month}{(Optional) A numeric vector of months to limit the results.}

\item{day}{(Optional) A numeric vector of days to limit the results.}

\item{period}{A numeric vector specifying the minimum and maximum duration (days) of events to include.  Can be a single value or a vector of length 2 (e.g., period = c(10, 30)). Maximum value is 180 days.}
}
\value{
A data frame containing the filtered oceanographic connectivity events.
}
\description{
\code{getConnectivityEvents} This function extracts a subset of oceanographic connectivity events from the main database ('oceanographicConnectivity') based on user-specified criteria including hexagon IDs, time periods (year, month, day), and the duration of events.
}
\examples{
\dontrun{
# Example: Filter by a single hexagon and year range
my_events <- getConnectivityEvents(hexagonID = "131", year = 2010:2012) 

# Example: Filter by multiple hexagons, month, and duration
my_events <- getConnectivityEvents(hexagonID = c("131", "132"), 
                                 month = 6, 
                                 period = c(15, 45)) 
}

}
