% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataBase.R
\name{getDataBase}
\alias{getDataBase}
\title{Get connectivity database from on-line repository}
\usage{
getDataBase(myFolder = "./Database", overwrite = FALSE)
}
\arguments{
\item{myFolder}{the folder path where the data files will be stored. Default is "./Database".}

\item{overwrite}{a logical value indicating whether to overwrite existing data files. Default is FALSE.}
}
\description{
\code{getDataBase} This function downloads and loads the required data for the coastal oceanographic connectivity analysis.
}
