% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDatasets.R
\docType{data}
\name{hexagonCells}
\alias{hexagonCells}
\title{Hexagons used in biophysical modelling}
\format{
An object of class ...
}
\usage{
data(hexagonCells)
}
\description{
A spatial object of class sp containing the hexagons used in biophysical modelling to estimate oceanographc connectivity.
}
