% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHexagons.R
\name{loadHexagons}
\alias{loadHexagons}
\title{Load R Data}
\usage{
loadHexagons(fileName = NULL)
}
\arguments{
\item{fileName}{The filepath of the sf file to load.}
}
\value{
The loaded data object from the sf file.'
}
\description{
\code{loadHexagons} A function to load the hexagons that define the source and sink locations of connectivity.
}
\examples{
\dontrun{
# Example with default hexagons
hexagonCells <- loadHexagons()

# Example with a local file
hexagonCells <- loadHexagons(fileName="data/my_hexagons.sf")
}

#' @importFrom sf read_sf

}
