% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Load R Data}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{The filepath or URL of the RData file to load.}
}
\value{
The loaded data object from the RData file.'
}
\description{
\code{loadRData} A function to load RData files from either a local filepath or a web-accessible URL. The function handles downloading from URLs if necessary.
}
\examples{
\dontrun{
# Example with a local file
mydata <- loadRData("data/my_data.RData")

# Example with a URL
remote_data <- loadRData("https://mywebsite.com/data/ocean_data.RData")
}

}
