% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapConnectivity.R
\name{mapConnectivity}
\alias{mapConnectivity}
\title{Visualize Pairwise Connectivity on a Map}
\usage{
mapConnectivity(connectivityPairs = NULL, print = FALSE)
}
\arguments{
\item{connectivityPairs}{A data frame containing pairwise connectivity results. It should have the following columns:
\itemize{
\item 'From' (character): The name or ID of the "from" region.
\item 'To' (character): The name or ID of the "to" region.
\item 'FromHexagon' (character): The hexagon ID of the starting point of the connection.
\item 'ToHexagon' (character): The hexagon ID of the ending point of the connection.
\item 'Value' (numeric): The connectivity value, used to determine the visual representation (e.g., line thickness or color) of the connection.
}}

\item{print}{(Optional) A logical flag. If 'TRUE', the map is displayed directly. Defaults to 'FALSE'.}
}
\value{
A list containing:
\itemize{
\item 'mappingData': The input 'connectivityPairs' data frame, augmented with longitude ('hexagonsFromLon', 'hexagonsToLon') and latitude ('hexagonsFromLat', 'hexagonsToLat') coordinates for the hexagons.
\item 'lineConnections': An 'sf' object representing the line connections between hexagons, with the 'Value' column used for visualization.
\item 'hexagonCells': An 'sf' object containing the hexagon polygons involved in the connections.
}
}
\description{
\code{mapConnectivity} This function creates a map visualization of pairwise oceanographic connectivity. It takes the output of the 'getPairwiseConnectivity' function (or a similar data frame) and plots connections between hexagon locations on a map, with the thickness or color of the connections representing the strength of the connectivity.
}
\examples{
\dontrun{
# Assuming you have the results from 'getPairwiseConnectivity' stored in 'pw_results'
map_data <- mapConnectivity(connectivityPairs = pw_results$connectivityPairs)

# Display the map:
mapConnectivity(connectivityPairs = pw_results$connectivityPairs, print = TRUE)
}

}
