% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDatasets.R
\docType{data}
\name{rastData}
\alias{rastData}
\title{My Data 3}
\format{
An object of class ...
}
\usage{
data(rastData)
}
\description{
A brief description of the first raster data, including its source, spatial resolution, extent, and any other relevant information.
}
