% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDatasets.R
\docType{data}
\name{referenceTable}
\alias{referenceTable}
\title{My Data 4}
\format{
An object of class ...
}
\usage{
data(referenceTable)
}
\description{
A brief description of the first raster data, including its source, spatial resolution, extent, and any other relevant information.
}
