% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDatasets.R
\docType{data}
\name{sfData}
\alias{sfData}
\title{My Data 5}
\format{
An object of class ...
}
\usage{
data(sfData)
}
\description{
A brief description of the first raster data, including its source, spatial resolution, extent, and any other relevant information.
}
