% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDatasets.R
\docType{data}
\name{summaryBM}
\alias{summaryBM}
\title{My Data 6}
\format{
An object of class ...
}
\usage{
data(summaryBM)
}
\description{
A brief description of the first raster data, including its source, spatial resolution, extent, and any other relevant information.
}
