% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnectivityEvents.R
\name{getConnectivityEvents}
\alias{getConnectivityEvents}
\title{Get connectivity events based on specified criteria}
\usage{
getConnectivityEvents(
  hexagonID,
  year = NULL,
  month = NULL,
  day = NULL,
  period = 180
)
}
\arguments{
\item{hexagonID}{a vector of hexagon IDs.}

\item{year}{a vector of years to filter the events by. Default is NULL.}

\item{month}{a vector of months to filter the events by. Default is NULL.}

\item{day}{a vector of days to filter the events by. Default is NULL.}

\item{period}{A numeric vector with one or two elements representing the time period in which the events occurred. Default is 180.
If a single value is provided, events with travel time less than or equal to the period will be returned.
If two values are provided, events with travel time between the two values (inclusive) will be returned.}
}
\value{
A subset of the oceanographicConnectivity database containing the selected connectivity events.
}
\description{
\code{getConnectivityEvents} This function retrieves connectivity events from the database based on the specified criteria.
}
