% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataBase.R
\name{getDataBase}
\alias{getDataBase}
\title{Retrieve and Summarize the Oceanographic Connectivity Database}
\usage{
getDataBase(myFolder = "./Database", overwrite = FALSE)
}
\arguments{
\item{myFolder}{(Optional) A character string specifying the local folder where the database will be stored. Defaults to "./Database".}

\item{overwrite}{(Optional) A logical flag. If 'TRUE', the existing database will be overwritten with a fresh download. Defaults to 'FALSE'.}
}
\value{
An object containing the connectivity database.
}
\description{
\code{getDataBase} This function manages the retrieval and loading of the oceanographic connectivity database. It can download the database
from an online repository if it does not exist locally or if the 'overwrite' flag is set. The function also provides a summary of the database's key characteristics.
}
\examples{
\dontrun{
# Download and load the database into the default folder
connectivity <- getDataBase()

# Download the database to a custom folder and force overwrite 
connectivity <- getDataBase(myFolder = "my_data_folder", overwrite = TRUE)
}

}
