% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHexagonID.R
\name{getHexagonID}
\alias{getHexagonID}
\title{Get the hexagon IDs based on a given spatial object and detail level}
\usage{
getHexagonID(obj, level = "extent", buffer = 0, print = FALSE)
}
\arguments{
\item{obj}{a spatial object of class (1) a matrix or data.frame with Longitude and Latitude coordinates (WGS84), (2) a numeric vector with c(minLon,maxLon,minLat,maxLat), (3) a polygon (sf) or (4) raster (SpatRaster) layer.}

\item{level}{the level of detail. Can be "extent" or "site". For the "extent" level, the function returns the hexagon IDs that intersect with the extent of the spatial object. For the "site" level, the function returns the hexagon IDs that are closest to the points or polygons of the spatial object. Default is "extent".}

\item{buffer}{a value defining the buffer distance (in decimal degrees) to expand the extent of the spatial object. Default is 0.}

\item{print}{a logical value indicating whether to print the hexagon sites. Default is FALSE.}
}
\description{
\code{getHexagonID} This function retrieves the hexagon IDs (i.e., coastal sites) given spatial object and detail level.
}
