% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPairwiseConnectivity.R
\name{getPairwiseConnectivity}
\alias{getPairwiseConnectivity}
\title{Calculate Pairwise Connectivity Estimates}
\usage{
getPairwiseConnectivity(
  connectivityEvents = NULL,
  hexagonIDFrom = NULL,
  hexagonIDTo = NULL,
  connType = "Forward",
  value = "Probability",
  steppingStone = FALSE,
  nStepStones = NULL,
  parallelCores = NULL
)
}
\arguments{
\item{connectivityEvents}{A data frame containing connectivity events.}

\item{hexagonIDFrom}{A list where each element is a vector of hexagon IDs representing 'from' regions.}

\item{hexagonIDTo}{(Optional) A list mirroring the structure of 'hexagonalIDFrom', specifying 'to' regions. If not provided, pairwise connectivity will be calculated for all pairs within 'hexagonIDFrom'.}

\item{connType}{A string specifying the type of connectivity direction. Options are "Forward" or "Backward". Defaults to "Forward".}

\item{value}{A string specifying the connectivity metric. Options are "Probability", "Events", or "Time". Defaults to "Probability".}

\item{steppingStone}{(Optional) A logical flag indicating whether stepping-stone connectivity should be considered. Defaults to FALSE.}

\item{nStepStones}{(Optional) If 'steppingStone' is TRUE, an integer specifying the maximum number of steps in a stepping-stone path. Defaults to unlimited.}

\item{parallelCores}{(Optional) The number of parallel cores to use for stepping-stone computations. If not provided, it defaults to available cores - 1.}
}
\value{
A list containing:
\itemize{
\item \strong{connectivityPairs:} A data frame of pairwise connectivity results (from, to, value, steps)
\item \strong{connectivityMatrix:} A matrix where rows are 'from' regions and columns are 'to' regions, containing connectivity values.
\item \strong{connectivityGraph:} An 'igraph' object representing the connectivity network
\item \strong{sitesConnected:} A vector of 'to' hexagon names that have connections
\item \strong{sitesNotConnected:} A vector of  'to' hexagon names with no connections, or NA if all are connected.
}
}
\description{
\code{getPairwiseConnectivity} This function calculates pairwise measures of oceanographic connectivity between sets of hexagons, considering various estimation metrics (probability, events, time),  forward or backward connectivity, and optional stepping stone calculations.
}
\examples{
\dontrun{
# Example: Basic connectivity (from hexagon 1 to 2)
pw_results <- getPairwiseConnectivity(connectivityEvents = my_events,
                                      hexagonIDFrom = list(c("1"), c("2")))

# Example: Stepping-stone connectivity, up to 3 steps, with parallel processing
pw_results <- getPairwiseConnectivity(connectivityEvents = my_events,
                                     hexagonIDFrom = list(c("1")), 
                                     hexagonIDTo =  list(c("2")), 
                                     steppingStone = TRUE, 
                                     nStepStones = 3,
                                     parallelCores = 2) 
}

}
