% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPairwiseConnectivity.R
\name{getPairwiseConnectivity}
\alias{getPairwiseConnectivity}
\title{Calculate pairwise connectivity}
\usage{
getPairwiseConnectivity(
  connectivityEvents = NULL,
  hexagonIDFrom = NULL,
  hexagonIDTo = NULL,
  connType = "Forward",
  value = "Probability",
  steppingStone = FALSE,
  nStepStones = NULL,
  parallelCores = NULL
)
}
\arguments{
\item{connectivityEvents}{A data frame containing all connectivity events.}

\item{hexagonIDFrom}{A numeric vector specifying the hexagon IDs from which connectivity is calculated.}

\item{hexagonIDTo}{A numeric vector specifying the hexagon IDs to which connectivity is calculated. Default is NULL.}

\item{connType}{A character string specifying the type of connectivity ("Forward" or "Backward"). Default is "Forward".}

\item{value}{A character string specifying the value to be calculated ("Probability" or "Time"). Default is "Probability".}

\item{steppingStone}{A logical value indicating whether to calculate stepping stone connectivity. Default is FALSE and will calculate direct connectivity.}

\item{nStepStones}{An integer specifying the maximum number of stepping stones to consider (if steppingStone is TRUE).}

\item{parallelCores}{An integer specifying the number of parallel cores to use in the calculation of stepping stone connectivity (if steppingStone is TRUE). Default is NULL and will use all available cores - 1.}
}
\value{
A list containing the pairwise connectivity information, including connectivity pairs, connectivity matrix, connectivity graph, sites connected, and sites not connected.
}
\description{
This function calculates pairwise connectivity between hexagon cells based on connectivity events.
}
\examples{
# Example usage of getPairwiseConnectivity function
connectivity <- getPairwiseConnectivity(connectivityEvents, hexagonIDFrom, hexagonIDTo, connType = "Forward", value = "Probability", steppingStone = FALSE)

}
