% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Load R Data}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{the name of the R data file to be loaded.}
}
\description{
\code{loadRData} This function loads an R data file and returns the objects stored in it.
}
