% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapConnectivity.R
\name{mapConnectivity}
\alias{mapConnectivity}
\title{Map connectivity between pairs of sites}
\usage{
mapConnectivity(connectivityPairs = NULL, obj = NULL, print = FALSE)
}
\arguments{
\item{connectivityPairs}{a data.frame containing probability / time estimates of connectivity between pairs of sites.}

\item{obj}{a data.frame containing the coordinates of the sites.}

\item{print}{a logical value indicating whether to print the mapped pairwise connectivity. Default is FALSE.}
}
\value{
a data.frame matching the pairwise connectivity data.frame with the coordinates of sites and a polygon of class sf with line connections between pairs of sites.
}
\description{
This function maps the connectivity between pairs of sites based on pairwise connectivity estimates and coordinate sites.
}
