% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapConnectivity.R
\name{mapConnectivity}
\alias{mapConnectivity}
\title{Visualize Pairwise Connectivity on a Map}
\usage{
mapConnectivity(connectivityPairs = NULL, print = FALSE)
}
\arguments{
\item{connectivityPairs}{A data frame containing pairwise connectivity results, ideally generated by the 'getPairwiseConnectivity' function. Should include columns: 'From', 'To', 'FromHexagon', 'ToHexagon', and 'Value'.}

\item{print}{(Optional) A logical flag. If TRUE, displays the map visualization directly. Defaults to FALSE.}
}
\value{
A list containing:
\itemize{
\item \strong{mappingData:} The input 'connectivityPairs' data frame, augmented with longitude and latitude coordinates.
\item \strong{lineConnections:} An 'sf' object representing the connections to be plotted.
\item \strong{hexagonCells:} An 'sf' object of hexagons involved in the visualized connections.
}
}
\description{
This function takes pairwise connectivity data and generates a map visualization highlighting the connections and their associated strengths. It requires external spatial data (\code{referenceTable}, \code{hexagonCells}).
}
\examples{
\dontrun{
# Assuming you have results from 'getPairwiseConnectivity' stored in 'pw_results'
map_data <- mapConnectivity(connectivityPairs = pw_results$connectivityPairs)

# To display the map:
mapConnectivity(connectivityPairs = pw_results$connectivityPairs, print = TRUE) 
}

}
